This README.md file describes the script run_analysis.R
describes the script run_analysis.R used to process the raw data
contained in the data set as described in the file. The purpose of the
script run_analysis.R is to demonstrate ability to collect, work with,
and clean a data set. The goal is to prepare tidy data that can be used
for later analysis.